requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
requests>=2.31.0
google-generativeai>=0.3.0
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
python-multipart>=0.0.6